Move KernelInterface backend tests to their own testsuite
