+++ lib/KernelInterface/test/runtests.jl
@@ -1,3 +1,5 @@
+# These are the standalone tests for KernelInterface
+
 using KernelInterface
 using Aqua
 using Test
