+++ test/testsuite.jl
@@ -27,7 +27,6 @@ end
 
 
 include("test.jl")
-include("interface.jl")
 include("hostinterface.jl")
 include("localmem.jl")
 include("private.jl")
@@ -49,10 +48,6 @@ function testsuite(backend, backend_str, backend_mod, AT, DAT; skip_tests = Set{
 
     @conditional_testset "SpecialFunctions" skip_tests begin
         specialfunctions_testsuite(backend)
-    end
-
-    @conditional_testset "Interface" skip_tests begin
-        interface_testsuite(backend, AT)
     end
 
     @conditional_testset "HostInterface" skip_tests begin
